[WL7-150]refactor:즐겨찾기 페이지 리팩토링
