+++ src/components/bookmark/index.ts
@@ -1,0 +1,3 @@
+// 북마크 관련 컴포넌트들
+export { default as BookmarkHeader } from './BookmarkHeader';
+export { default as BookmarkList } from './BookmarkList';
