+++ src/components/common/StoreTypeIcon.tsx
@@ -177,43 +177,3 @@ const StoreTypeIcon: React.FC<StoreTypeIconProps> = ({
 };
 
 export default StoreTypeIcon;
-
-/*
-사용법 (백엔드 API 스펙에 맞춤)
-
-기본 사용법 (매장 모드, 네모 모양):
-<StoreTypeIcon category="CAFE" storeClass="FRANCHISE" />
-
-팝업스토어 카테고리:
-<StoreTypeIcon category="POPUP" storeClass="BASIC" />
-
-기본 매장:
-<StoreTypeIcon category="FOOD" storeClass="BASIC" />
-
-이벤트 포함:
-<StoreTypeIcon category="FOOD" storeClass="LOCAL" event="GENERAL" />
-
-원형 모양:
-<StoreTypeIcon category="CAFE" storeClass="FRANCHISE" shape="circle" />
-
-통계 모드 (카테고리별 색상):
-<StoreTypeIcon category="CAFE" storeClass="FRANCHISE" mode="statistics" />
-
-Props:
-- category: 카테고리 (FOOD, ACTIVITY, EDUCATION, CULTURE, BAKERY, LIFE, SHOPPING, CAFE, BEAUTY, POPUP)
-- storeClass: 매장 구분 (LOCAL, FRANCHISE, BASIC)
-- event: 이벤트 타입 (NONE, GENERAL, REQUIRE) - 선택적, 기본값: NONE
-- size: 크기 (기본값: 50)
-- className: 추가 CSS 클래스
-- mode: 색상 모드 (store, statistics) - 기본값: store
-- shape: 모양 (square, circle) - 기본값: square
-
-색상 규칙 (store 모드):
-- 이벤트 우선순위:
-  * REQUIRE 이벤트: pink-400
-  * GENERAL 이벤트: 핑크색 (text-primary) - 이벤트 매장
-- 매장 구분:
-  * FRANCHISE/BASIC: 주황색 (text-orange-500)
-  * LOCAL: 파란색 (text-blue-500) - 소상공인
-- statistics 모드: 카테고리별 고유 색상 (POPUP 포함)
-*/
