+++ src/hooks/index.ts
@@ -3,3 +3,6 @@ export * from './my';
 
 // 이용내역 관련 훅들 (현재 사용되지 않으므로 주석 처리)
 // export { useUsageHistory } from './usagehistory/useUsageHistory';
+
+// 북마크 관련 훅들
+export * from './bookmark';
